requirements.txt
Django==5.1.5
django-widget-tweaks==1.5.0
factory_boy==3.3.1
